+++ terezinha-farm/api/src/schemaComposer.ts
@@ -1,5 +1,5 @@
 import './modules/Farm/QueryFarmsComposer';
 
-import './modules/User/composer';
+//import './modules/User/composer';
 
 export { schemaComposer } from '@ttoss/graphql-api';