+++ docs/storybook/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.16](https://github.com/ttoss/ttoss/compare/@docs/storybook@2.0.15...@docs/storybook@2.0.16) (2024-01-12)
+
+**Note:** Version bump only for package @docs/storybook
+
+## [2.0.15](https://github.com/ttoss/ttoss/compare/@docs/storybook@2.0.14...@docs/storybook@2.0.15) (2024-01-12)
+
+**Note:** Version bump only for package @docs/storybook
+
 ## [2.0.14](https://github.com/ttoss/ttoss/compare/@docs/storybook@2.0.13...@docs/storybook@2.0.14) (2024-01-10)
 
 **Note:** Version bump only for package @docs/storybook