+++ packages/carlin/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.31.9](https://github.com/ttoss/ttoss/compare/carlin@1.31.8...carlin@1.31.9) (2024-01-12)
+
+**Note:** Version bump only for package carlin
+
 ## [1.31.8](https://github.com/ttoss/ttoss/compare/carlin@1.31.7...carlin@1.31.8) (2024-01-02)
 
 ### Bug Fixes