+++ terezinha-farm/api/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.1](https://github.com/ttoss/ttoss/compare/@terezinha-farm/api@0.13.0...@terezinha-farm/api@0.13.1) (2024-01-12)
+
+**Note:** Version bump only for package @terezinha-farm/api
+
+# [0.13.0](https://github.com/ttoss/ttoss/compare/@terezinha-farm/api@0.12.0...@terezinha-farm/api@0.13.0) (2024-01-12)
+
+### Features
+
+- create tests to graphql server ([#474](https://github.com/ttoss/ttoss/issues/474)) ([86d5a89](https://github.com/ttoss/ttoss/commit/86d5a89491f52abec3564a72da50fde488274b2b))
+
 # [0.12.0](https://github.com/ttoss/ttoss/compare/@terezinha-farm/api@0.11.2...@terezinha-farm/api@0.12.0) (2024-01-03)
 
 ### Features